added BaseIntrospector and reorganized connections
