+++ app/__test__/data/specs/connection/SqliteIntrospector.spec.ts
@@ -1,8 +1,7 @@
-import { beforeEach, describe, test, expect } from "bun:test";
-import { SqliteIntrospector } from "data/connection/SqliteIntrospector";
-import { getDummyConnection, getDummyDatabase } from "../../helper";
-import { Kysely, ParseJSONResultsPlugin, SqliteDialect } from "kysely";
-import { _jsonp } from "core/utils";
+import { describe, expect, test } from "bun:test";
+import { SqliteIntrospector } from "data/connection";
+import { getDummyDatabase } from "../../helper";
+import { Kysely, SqliteDialect } from "kysely";
 
 function create() {
    const database = getDummyDatabase().dummyDb;
